pyproject.toml
[project]
name = "urlshort"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dj-database-url>=2.3.0",
    "django-cors-headers>=4.7.0",
    "django-ratelimit>=4.1.0",
    "django-redis>=5.4.0",
    "djangorestframework>=3.16.0",
    "drf-yasg>=1.21.10",
    "psycopg2-binary>=2.9.10",
    "whitenoise>=6.9.0",
]
